fix(review): isolate Codex readiness adapter
